+++ CHANGELOG.md
@@ -12,6 +12,7 @@
     * [Added grid context param to columns render method](https://github.com/softindex/uikernel/pull/269)
     * [Added `multipart/form-data` encoding to GridXHRModel](https://github.com/softindex/uikernel/pull/270)
     * [Added toCSV export](https://github.com/softindex/uikernel/pull/261)
+    * [Added logger param to GridExpressApi](https://github.com/softindex/uikernel/pull/271)
     * Added `recordId` property to GridModel.prototype.isValidRecord
 
   * Form